refactor: 패키지 구조 요구사항에 맞게 변경
